This should be the first version of Tfit capable of working with a ParamWrapper instead of a fixed params map.
Why is this change useful?
By abstracting parameters, it's now possible to rework and improve the overall inteface of Tfit without
breaking the entire program. Before this change, a large number of functions would directly access parameters
by their human visibile names. This means that, say, if we were to replace the "-i" parameter with "-forward",
then every single instance in which a forward strand file is desired would have to be updated.